A rewrite of the recaptcha verify API and move of the templating into a separate module.
